Properly cast to ULONGLONG before the multiplication to avoid an integer overflow.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@25311 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/CommitDlg.cpp
@@ -1014,7 +1014,7 @@ void CCommitDlg::GetAutocompletionList()
     std::map<CString, CString> mapRegex;
     CString sRegexFile = CPathUtils::GetAppDirectory();
     CRegDWORD regtimeout = CRegDWORD(L"Software\\TortoiseSVN\\AutocompleteParseTimeout", 5);
-    ULONGLONG timeoutvalue = regtimeout * 1000;
+    ULONGLONG timeoutvalue = ULONGLONG(DWORD(regtimeout)) * 1000UL;
     if (timeoutvalue == 0)
         return;
     sRegexFile += L"autolist.txt";
